+++ src/encoded/tests/test_types_protocol.py
@@ -1,7 +1,8 @@
-import datetime
+# import datetime
 import pytest
 
-from ..util import utc_today_str
+from dcicutils.misc_utils import utc_today_str
+# from ..util import utc_today_str
 
 pytestmark = [pytest.mark.setone, pytest.mark.working, pytest.mark.schema]
